+++ src/plugins/storage-denokv/rx-storage-instance-denokv.ts
@@ -24,7 +24,7 @@ import type { DenoKVIndexMeta, DenoKVSettings, DenoKVStorageInternals } from './
 import { RxStorageDenoKV } from './index.ts';
 import { CLEANUP_INDEX, DENOKV_DOCUMENT_ROOT_PATH, RX_STORAGE_NAME_DENOKV, commitWithRetry, getDenoGlobal, getDenoKVIndexName } from "./denokv-helper.ts";
 import { getIndexableStringMonad, getStartIndexStringFromLowerBound } from "../../custom-index.ts";
-import { appendToArray, batchArray, lastOfArray, toArray } from "../utils/utils-array.ts";
+import { batchArray, lastOfArray, toArray } from "../utils/utils-array.ts";
 import { ensureNotFalsy } from "../utils/utils-other.ts";
 import { categorizeBulkWriteRows } from "../../rx-storage-helper.ts";
 import { now } from "../utils/utils-time.ts";
@@ -189,7 +189,7 @@ export class RxStorageInstanceDenoKV<RxDocType> implements RxStorageInstance<
                     }
                 }
                 if (txResult && txResult.ok) {
-                    appendToArray(ret.error, categorized.errors);
+                    ret.error = ret.error.concat(categorized.errors);
                     if (categorized.eventBulk.events.length > 0) {
                         const lastState = ensureNotFalsy(categorized.newestRow).document;
                         categorized.eventBulk.checkpoint = {