+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### 17.0.0-beta.3 (7 January 2026)
 
+- Replace `appendToArray()` with `Array.concat()` which is nowadays optimized by browsers and faster.
 - ADD  `reactivity-angular` package.
 - CHANGE moved `reactivity-vue` and `reactivity-preact-signals` from premium to core.
 - ADD `databaseNamePrefix` option to premium SQLite RxStorage